+++ UniversityCourseSelection/src/main/java/com/capgemini/UniversityCourseSelection/controllers/ApplicantController.java
@@ -1,8 +1,8 @@
 package com.capgemini.UniversityCourseSelection.controllers;
 
-
 import com.capgemini.UniversityCourseSelection.entities.Applicant;
-import com.capgemini.UniversityCourseSelection.repo.ILoginRepository;
+import com.capgemini.UniversityCourseSelection.exception.NotFoundException;
+import com.capgemini.UniversityCourseSelection.exception.NotLoggedInException;
 import com.capgemini.UniversityCourseSelection.services.IApplicantService;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,12 +15,13 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 import java.util.Optional;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 
 @RestController
 @RequestMapping("/applicant")
@@ -28,9 +29,21 @@ public class ApplicantController {
 
 	@Autowired
 	private IApplicantService service;
-//	
-//	@Autowired
-//	private ILoginRepository repo;
+
+	public boolean checkSession(HttpServletRequest request, String type) {
+		HttpSession session = request.getSession();
+
+		boolean validLogin = true;
+		if (session.isNew())
+			return false;
+		if (session.getAttribute(type) == null)
+			return false;
+		int userId = (int) session.getAttribute(type);
+		if (userId == 0)
+			validLogin = false;
+
+		return validLogin;
+	}
 
 	@PostMapping("/apply")
 	public ResponseEntity<Applicant> applyForCourse(@RequestBody Applicant applicant) {
@@ -42,7 +55,23 @@ public class ApplicantController {
 	}
 
 	@PutMapping("/update")
-	public ResponseEntity<Applicant> updateApplication(@RequestBody Applicant applicant) {
+	public ResponseEntity<Applicant> updateApplication(@RequestBody Applicant applicant,HttpServletRequest request) {
+		
+		boolean valid=checkSession(request,"applicant");
+		String host = String.valueOf(request.getServerPort());
+		if(!valid) {
+			throw new NotLoggedInException("Please Login to update details, click " + host
+					+ "/login/applicant to login");
+		}
+		
+		if (applicant == null || applicant.getApplicantId() == null) {
+			throw new NotFoundException("Applicant or Id can't be null!");
+		}
+		HttpSession session=request.getSession();
+		if(applicant.getApplicantId()!=(int)session.getAttribute("applicant")){
+			throw new NotLoggedInException("You can only update your own details");
+		}
+		
 
 		Applicant temp = service.updateApplicant(applicant);
 
@@ -51,37 +80,66 @@ public class ApplicantController {
 	}
 
 	@DeleteMapping("/delete")
-	public ResponseEntity<Applicant> deleteApplication(@RequestBody Applicant applicant) {
+	public ResponseEntity<Applicant> deleteApplication(@RequestBody Applicant applicant, HttpServletRequest request) {
+		boolean valid = checkSession(request, "commitee");
+		String host = String.valueOf(request.getServerPort());
+		
+		if (!valid) {
+			throw new NotLoggedInException(
+					"Accessible to commitee members only. If you are a registered commitee member, click " + host
+							+ "/login/commitee to login.");
+
+		}
+		if (applicant == null || applicant.getApplicantId() == null) {
+			throw new NotFoundException("Applicant or Id can't be null!");
+		}
 		Applicant temp = service.deleteApplicant(applicant);
 		return new ResponseEntity<>(temp, HttpStatus.OK);
 
 	}
 
-	
 	@GetMapping("/get/{id}")
-	public ResponseEntity<Applicant> getById(@PathVariable("id") Integer id){
-		Optional<Applicant> temp= service.viewApplicant(id);
-		if(temp.isEmpty()) {
-			return new ResponseEntity("No such user",HttpStatus.NOT_FOUND);
-					
+	public ResponseEntity<Applicant> getById(@PathVariable int id, HttpServletRequest request) {
+
+		boolean valid = checkSession(request, "applicant") || checkSession(request, "commitee")
+				|| checkSession(request, "staffMember");
+		String host = String.valueOf(request.getServerPort());
+		if (!valid) {
+			throw new NotLoggedInException(
+					"Kindly login to view your details.  click " + host + "/login/applicant to login.");
+
 		}
-		return new ResponseEntity<>(temp.get(),HttpStatus.OK);
+		HttpSession session = request.getSession();
+
+		Optional<Applicant> temp = service.viewApplicant(id);
+		if (temp.isEmpty()) {
+			throw new NotFoundException("No user with given Id is present");
+
+		}
+		if (checkSession(request, "applicant")) {
+			int attId = (int) session.getAttribute("applicant");
+			if (attId != id) {
+				temp.get().setPassword("******");
+			}
+		} else {
+			temp.get().setPassword("******");
+		}
+		return new ResponseEntity<>(temp.get(), HttpStatus.OK);
 	}
-	
+
 	@GetMapping("/getAll/{status}")
-	public ResponseEntity<List<Applicant>> getAllApplicants(@PathVariable int status){
-		
-		List<Applicant> res= service.viewAllApplicantsByStatus(status);
-		return new ResponseEntity<>(res,HttpStatus.OK);
-		
+	public ResponseEntity<List<Applicant>> getAllApplicants(@PathVariable int status, HttpServletRequest request) {
+		boolean valid = checkSession(request, "commitee");
+		String host = String.valueOf(request.getServerPort());
+		if (!valid) {
+			throw new NotLoggedInException(
+					"Accessible to commitee members only. If you are a registered commitee member, click " + host
+							+ "/login/commitee to login.");
+
+		}
+
+		List<Applicant> res = service.viewAllApplicantsByStatus(status);
+		return new ResponseEntity<>(res, HttpStatus.OK);
 	}
-	
-//	@PostMapping("/loginTest")
-//	@ResponseBody
-//	public ResponseEntity<String> loginTest(@RequestParam Integer id, @RequestParam String password){
-//		if(repo.verifyApplicantCredentials(id, password)) {
-//			return new ResponseEntity<String>("You are logged in",HttpStatus.OK);
-//		}
-//		return new ResponseEntity<>("Enter valid credentials please",HttpStatus.FORBIDDEN);
-//	}
+
 }
